Add Geolocation Redirection for specific ip
This patch redirects you to the correct library site if it matches your
location with the saved location of a library. Hardcoded lat and long
values are included as this is a proof of concept for LibrariesWest
demo.
Also saves redirect data in browser session cookie to prevent redirect
loops