Add Manual redirection to specific library sites with a dropdown box
This patch adds the ability to swap to a specific library site with a
dropdown, the library list is currently hard coded and not dynamic as
this is a proof of concept for librariesWest.
Should also only appear on west.aspendiscovery and east.aspendiscovery
as to not disrupt other sales teams members workflow.